dragonboard: temporarily copy gbm_gralloc into dragonboard
While the upstream status of gbm_gralloc patches is sorted out,
move it to device/linaro/dragonboard - right now pointing to John's
version [1], which includes the libdrm gralloc handle changes.

Though the current upstream for this is Rob Herring's tree [2],
the ultimate goal for this is to reside in mesa, and not in dragonboard
project or Rob's.

[1]: https://github.com/johnstultz-work/gbm_gralloc/commits/drm_hwc
     SHA: 528cfc45195495560de5348677eeee6d76d02dbf
[2]: https://github.com/robherring/gbm_gralloc

Change-Id: I4a38348ded85c91e2283328465af04524f8c36ff
Signed-off-by: Sumit Semwal <sumit.semwal@linaro.org>