Add factory-image generation scripts
Add scripts to generate the db410c32_only and db820c factory images
and include a README to explain how to use the scripts to properly
generate the factory images.

"fastboot update <zip>" is not flashing cache and userdata partitions
for some reason. So we don't do "fastboot update" and instead fastboot
flash the AOSP images separately. #FIXME

Change-Id: I442cffcf62cc0a343802bd602c828e2fd67f7859
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>